# sql/catalog: reduce NameKey allocations