--- a/pkg/sql/catalog/nstree/catalog.go
+++ b/pkg/sql/catalog/nstree/catalog.go
@@ -170,8 +170,8 @@ func (c Catalog) LookupZoneConfig(id descpb.ID) catalog.ZoneConfig {
 }
 
 // LookupNamespaceEntry looks up a descriptor ID by name.
-func (c Catalog) LookupNamespaceEntry(key catalog.NameKey) NamespaceEntry {
-	if !c.IsInitialized() || key == nil {
+func (c Catalog) LookupNamespaceEntry(key descpb.NameInfo) NamespaceEntry {
+	if !c.IsInitialized() {
 		return nil
 	}
 	e := c.byName.getByName(key.GetParentID(), key.GetParentSchemaID(), key.GetName())
@@ -257,7 +257,7 @@ func (c Catalog) Validate(
 // ValidateNamespaceEntry returns an error if the specified namespace entry
 // is invalid.
 func (c Catalog) ValidateNamespaceEntry(key catalog.NameKey) error {
-	ne := c.LookupNamespaceEntry(key)
+	ne := c.LookupNamespaceEntry(catalog.MakeNameInfo(key))
 	if ne == nil {
 		return errors.AssertionFailedf("invalid namespace entry")
 	}