--- a/pkg/sql/catalog/replication/reader_catalog.go
+++ b/pkg/sql/catalog/replication/reader_catalog.go
@@ -123,7 +123,8 @@ func SetupOrAdvanceStandbyReaderCatalog(
 					return nil
 				}
 				// Do not upsert entries if one already exists.
-				if entry := allExistingDescs.LookupNamespaceEntry(e); entry != nil && e.GetID() == entry.GetID() {
+				entry := allExistingDescs.LookupNamespaceEntry(catalog.MakeNameInfo(e))
+				if entry != nil && e.GetID() == entry.GetID() {
 					return nil
 				}
 				return errors.Wrapf(txn.Descriptors().UpsertNamespaceEntryToBatch(ctx, true, e, b), "namespace entry %v", e)