#### sql/catalog: change `LookupNamespaceEntry` parameter type

The `Catalog.LookupNamespaceEntry` method now accepts the concrete type
`descpb.NameInfo` instead of the `catalog.NameKey` interface. This,
along with future commits, will help reduce allocations.

Release note: None

#### sql/catalog: change `IsNameInCache` parameter type

The `CatalogReader.IsNameInCache` method now accepts the concrete type
`descpb.NameInfo` instead of the `catalog.NameKey` interface. This,
along with other commits, will help reduce allocations.

Release note: None

#### sql/catalog: change `(*Collection).isShadowedName` parameter type

The `(*Collection).isShadowedName` method now accepts the concrete type
`descpb.NameInfo` instead of the `catalog.NameKey` interface. This,
along with other commits, will help reduce allocations.

Informs #105867
Fixes #135905
Fixes #121300

Release note: None

#### sql/catalog: refactor `(*SystemDatabaseCache).lookupDescriptorID`

Now that `(*SystemDatabaseCache).lookupDescriptorID` accepts a
`descpb.NameInfo` parameter, it can access its exported field directly
instead of using getters.

Release note: None