@mgartner 2024-11-26
`BenchmarkEndToEnd/kv-read` shows the improvement:

```
name                                     old time/op    new time/op    delta
EndToEnd/kv-read/vectorize=on/Simple        422µs ±10%     422µs ± 4%    ~     (p=0.690 n=5+5)
EndToEnd/kv-read/vectorize=on/Prepared      272µs ± 6%     276µs ± 2%    ~     (p=0.841 n=5+5)
EndToEnd/kv-read/vectorize=off/Simple       415µs ± 2%     428µs ± 7%    ~     (p=0.690 n=5+5)
EndToEnd/kv-read/vectorize=off/Prepared     269µs ± 4%     271µs ± 6%    ~     (p=1.000 n=5+5)

name                                     old alloc/op   new alloc/op   delta
EndToEnd/kv-read/vectorize=on/Simple       36.2kB ± 1%    36.0kB ± 1%    ~     (p=0.343 n=4+4)
EndToEnd/kv-read/vectorize=on/Prepared     25.9kB ±16%    26.0kB ±17%    ~     (p=1.000 n=5+5)
EndToEnd/kv-read/vectorize=off/Simple      37.8kB ± 4%    38.1kB ±11%    ~     (p=0.421 n=5+5)
EndToEnd/kv-read/vectorize=off/Prepared    26.0kB ± 2%    25.8kB ± 0%    ~     (p=0.111 n=5+4)

name                                     old allocs/op  new allocs/op  delta
EndToEnd/kv-read/vectorize=on/Simple          321 ± 0%       311 ± 0%  -3.19%  (p=0.029 n=4+4)
EndToEnd/kv-read/vectorize=on/Prepared        237 ±12%       233 ±13%    ~     (p=0.190 n=5+5)
EndToEnd/kv-read/vectorize=off/Simple         313 ± 0%       303 ± 0%  -3.19%  (p=0.029 n=4+4)
EndToEnd/kv-read/vectorize=off/Prepared       221 ± 0%       216 ± 0%  -2.09%  (p=0.000 n=5+4)
```