@rafiss 2024-11-26
does this close out https://github.com/cockroachdb/cockroach/issues/121300 and https://github.com/cockroachdb/cockroach/issues/135905?