@mgartner 2024-11-27
> does this close out https://github.com/cockroachdb/cockroach/issues/121300 and https://github.com/cockroachdb/cockroach/issues/135905?

Yes, it fixes both of them. I've added them to the PR description. Thanks for finding those!